backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
langchain>=0.2.0
langchain-google-genai>=1.0.6
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
pydantic>=2.7.0
python-dotenv>=1.0.1
chardet>=5.0.0
